Notifier skeleton + Cron + Crawler retries
Crawler tries N times to locate price by reloading page.
Notifier checks last 2 prices entries of product and if
price change is significant given the config params, it
logs notification (not ntfy send) for now.
Cron of each product is guaranteed to always be refilled
with future runtimes when it runs low.